+++ aiida/tools/data/array/kpoints/seekpath.py
@@ -12,7 +12,7 @@ from __future__ import print_function
 from __future__ import absolute_import
 
 from aiida.orm.nodes.data.array.kpoints import KpointsData
-from aiida.orm.nodes.data.parameter import ParameterData
+from aiida.orm.nodes.data.dict import Dict
 
 __all__ = ('check_seekpath_is_installed', 'get_explicit_kpoints_path', 'get_kpoints_path')
 
@@ -50,7 +50,7 @@ def get_explicit_kpoints_path(structure, parameters):
         - ``explicit_kpoints``: a KpointsData with the (explicit) kpoints
           (with labels set).
 
-        - ``parameters``: a ParameterData, whose content is
+        - ``parameters``: a Dict, whose content is
           the same dictionary as returned by the ``seekpath.get_explicit_k_path`` function
           (see `seekpath documentation <https://seekpath.readthedocs.io/>`_),
           except that:
@@ -105,7 +105,7 @@ def get_explicit_kpoints_path(structure, parameters):
     kpoints.set_cell_from_structure(primitive_structure)
     kpoints.set_kpoints(kpoints_abs, cartesian=True, labels=labels)
 
-    result['parameters'] = ParameterData(dict=rawdict)
+    result['parameters'] = Dict(dict=rawdict)
     result['explicit_kpoints'] = kpoints
     result['primitive_structure'] = primitive_structure
     result['conv_structure'] = conv_structure
@@ -133,7 +133,7 @@ def get_kpoints_path(structure, parameters):
 
     :return: A dictionary with three nodes:
 
-        - ``parameters``: a ParameterData, whose content is
+        - ``parameters``: a Dict, whose content is
           the same dictionary as returned by the ``seekpath.get_path`` function
           (see `seekpath documentation <https://seekpath.readthedocs.io/>`_),
           except that:
@@ -172,7 +172,7 @@ def get_kpoints_path(structure, parameters):
     primitive_tuple = (primitive_lattice, primitive_positions, primitive_types)
     primitive_structure = spglib_tuple_to_structure(primitive_tuple, kind_info, kinds)
 
-    result['parameters'] = ParameterData(dict=rawdict)
+    result['parameters'] = Dict(dict=rawdict)
     result['primitive_structure'] = primitive_structure
     result['conv_structure'] = conv_structure