+++ aiida/tools/dbexporters/tcod.py
@@ -18,7 +18,7 @@ from six.moves import range
 import io
 
 from aiida.engine import calcfunction
-from aiida.orm.nodes.data.parameter import ParameterData
+from aiida.orm.nodes.data.dict import Dict
 from aiida.plugins import DataFactory
 
 aiida_executable_name = '_aiidasubmit.sh'
@@ -624,16 +624,16 @@ def extend_with_cmdline_parameters(parser, expclass="Data"):
                         dest='reduce_symmetry',
                         help="Do not perform symmetry reduction.")
     parser.add_argument('--parameter-data', type=int, default=None,
-                        help="ID of the ParameterData to be exported "
+                        help="ID of the Dict to be exported "
                              "alongside the {} instance. "
                              "By default, if {} originates from "
-                             "a calculation with single ParameterData "
+                             "a calculation with single Dict "
                              "in the output, aforementioned "
-                             "ParameterData is picked automatically. "
+                             "Dict is picked automatically. "
                              "Instead, the option is used in the case "
                              "the calculation produces more than a "
                              "single instance of "
-                             "ParameterData.".format(expclass,expclass))
+                             "Dict.".format(expclass,expclass))
     parser.add_argument('--dump-aiida-database', action='store_true',
                         default=None,
                         dest='dump_aiida_database',
@@ -890,10 +890,10 @@ def add_metadata_inline(what, node, parameters, args):
 
     :param what: an original exported node.
     :param node: a :py:class:`aiida.orm.nodes.data.cif.CifData` instance.
-    :param parameters: a :py:class:`aiida.orm.nodes.data.parameter.ParameterData`
+    :param parameters: a :py:class:`aiida.orm.nodes.data.dict.Dict`
         instance, produced by the same calculation as the original exported
         node.
-    :param args: a :py:class:`aiida.orm.nodes.data.parameter.ParameterData`
+    :param args: a :py:class:`aiida.orm.nodes.data.dict.Dict`
         instance, containing parameters for the control of metadata
         collection and inclusion in the produced
         :py:class:`aiida.orm.nodes.data.cif.CifData`.
@@ -921,7 +921,7 @@ def add_metadata_inline(what, node, parameters, args):
     for loop in node.values[dataname].loops.values():
         loops[loop[0]] = loop
 
-    # Unpacking the kwargs from ParameterData
+    # Unpacking the kwargs from Dict
     kwargs = {}
     additional_tags = {}
     datablock_names = None
@@ -983,7 +983,7 @@ def export_cifnode(what, parameters=None, trajectory_index=None,
     supported in addition to :py:class:`aiida.orm.nodes.data.cif.CifData`.
 
     :param what: data node to be exported.
-    :param parameters: a :py:class:`aiida.orm.nodes.data.parameter.ParameterData`
+    :param parameters: a :py:class:`aiida.orm.nodes.data.dict.Dict`
         instance, produced by the same calculation as the original exported
         node.
     :param trajectory_index: a step to be converted and exported in case a
@@ -1010,21 +1010,21 @@ def export_cifnode(what, parameters=None, trajectory_index=None,
     CifData        = DataFactory('cif')
     StructureData  = DataFactory('structure')
     TrajectoryData = DataFactory('array.trajectory')
-    ParameterData  = DataFactory('parameter')
+    Dict  = DataFactory('dict')
 
     calc = _get_calculation(what)
 
     if parameters is not None:
-        if not isinstance(parameters, ParameterData):
+        if not isinstance(parameters, Dict):
             raise ValueError("Supplied parameters are not an "
-                             "instance of ParameterData")
+                             "instance of Dict")
     elif calc is not None:
-        params = calc.get_outgoing(node_class=ParameterData, link_type=LinkType.CREATE).all()
+        params = calc.get_outgoing(node_class=Dict, link_type=LinkType.CREATE).all()
         if len(params) == 1:
             parameters = params[0].node
         elif len(params) > 0:
             raise MultipleObjectsError("Calculation {} has more than "
-                                       "one ParameterData output, please "
+                                       "one Dict output, please "
                                        "specify which one to use with "
                                        "an option parameters='' when "
                                        "calling export_cif()".format(calc))
@@ -1055,14 +1055,14 @@ def export_cifnode(what, parameters=None, trajectory_index=None,
 
     # Addition of the metadata
 
-    args = ParameterData(dict=kwargs)
+    args = Dict(dict=kwargs)
     function_args = {'what': what, 'args': args, 'metadata': {'store_provenance': store}}
     if node != what:
         function_args['node'] = node
     if parameters is not None:
         function_args['parameters'] = parameters
     else:
-        function_args['parameters'] = ParameterData(dict={})
+        function_args['parameters'] = Dict(dict={})
     ret_dict = add_metadata_inline(**function_args)
 
     return ret_dict['cif']
@@ -1164,7 +1164,7 @@ def deposit(what, type, author_name=None, author_email=None, url=None,
         parameters['replace'] = True
     if message:
         parameters['log-message'] = str(message)
-    pd = ParameterData(dict=parameters)
+    pd = Dict(dict=parameters)
 
     calc.use_cif(cif)
     calc.use_parameters(pd)